+++ plugins/set_value/release_operation.cxx
@@ -1,4 +1,5 @@
 /*******************************************************************************
+ * Copyright (C) 2026 Commissariat a l'energie atomique et aux energies alternatives (CEA)
  * Copyright (C) 2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
@@ -23,28 +24,28 @@
  ******************************************************************************/
 
 #include <pdi/context.h>
+#include <pdi/logger.h>
 
 #include "release_operation.h"
 
 namespace set_value {
 
-Release_operation::Release_operation(PDI::Context& ctx, PC_tree_t release_value_node)
-	: Operation{ctx}
+Release_operation::Release_operation(PDI::Logger& logger, PC_tree_t release_value_node)
 {
 	size_t list_size = PDI::len(release_value_node);
-	context().logger().debug("Release operation count: {}", list_size);
+	logger.debug("Release operation count: {}", list_size);
 	for (int i = 0; i < list_size; i++) {
 		std::string data_to_release = PDI::to_string(PC_get(release_value_node, "[%d]", i));
-		context().logger().trace("\t {}: {}", i, data_to_release);
+		logger.trace("\t {}: {}", i, data_to_release);
 		m_data_to_release.emplace_back(data_to_release);
 	}
 }
 
-void Release_operation::execute()
+void Release_operation::execute(PDI::Logger& logger, PDI::Context& ctx)
 {
 	for (auto&& data_to_release: m_data_to_release) {
-		context().logger().trace("Calling {} data release", data_to_release);
-		context().desc(data_to_release).release();
+		logger.trace("Calling {} data release", data_to_release);
+		ctx.desc(data_to_release).release();
 	}
 }